Add files via upload
The Data is of an Online retailer which has just launched its first product.  For this Project, I am going to analyze and optimize marketing channels. And going to measure and test the website conversion performance, And use data to understand the impact when the company launches a new product. I will be making recommendations to steer the online retailer business, and see how the business evolves based on the analyses.